User better environment variable names
As pointed out in #2, the previous names were very generic, and it might conflict with other GitHub related variables
people have in their environments.

These names are now very explicit with the `TIL_RB` prefix, and we will introduce a better way, using a file under
`~/.config` later on to remove the need for environment variables.

Fixes #2